+++ examples/contrib/tls_passthrough.py
@@ -1,8 +1,5 @@
-# FIXME: This addon is currently not compatible with mitmproxy 7 and above.
-
 """
-This inline script allows conditional TLS Interception based
-on a user-defined strategy.
+This addon allows conditional TLS Interception based on a user-defined strategy.
 
 Example:
 
@@ -11,138 +8,103 @@ Example:
     1. curl --proxy http://localhost:8080 https://example.com --insecure
     // works - we'll also see the contents in mitmproxy
 
-    2. curl --proxy http://localhost:8080 https://example.com --insecure
-    // still works - we'll also see the contents in mitmproxy
-
-    3. curl --proxy http://localhost:8080 https://example.com
+    2. curl --proxy http://localhost:8080 https://example.com
     // fails with a certificate error, which we will also see in mitmproxy
 
-    4. curl --proxy http://localhost:8080 https://example.com
+    3. curl --proxy http://localhost:8080 https://example.com
     // works again, but mitmproxy does not intercept and we do *not* see the contents
-
-Authors: Maximilian Hils, Matthew Tuusberg
 """
 import collections
 import random
-
+from abc import ABC, abstractmethod
 from enum import Enum
 
-import mitmproxy
-from mitmproxy import ctx
-from mitmproxy.exceptions import TlsProtocolException
-from mitmproxy.proxy.protocol import TlsLayer, RawTCPLayer
+from mitmproxy import connection, ctx
+from mitmproxy.proxy.layers import tls
+from mitmproxy.utils import human
 
 
 class InterceptionResult(Enum):
-    success = True
-    failure = False
-    skipped = None
+    SUCCESS = 1
+    FAILURE = 2
+    SKIPPED = 3
 
 
-class _TlsStrategy:
-    """
-    Abstract base class for interception strategies.
-    """
-
+class TlsStrategy(ABC):
     def __init__(self):
         # A server_address -> interception results mapping
         self.history = collections.defaultdict(lambda: collections.deque(maxlen=200))
 
-    def should_intercept(self, server_address):
-        """
-        Returns:
-            True, if we should attempt to intercept the connection.
-            False, if we want to employ pass-through instead.
-        """
+    @abstractmethod
+    def should_intercept(self, server_address: connection.Address) -> bool:
         raise NotImplementedError()
 
     def record_success(self, server_address):
-        self.history[server_address].append(InterceptionResult.success)
+        self.history[server_address].append(InterceptionResult.SUCCESS)
 
     def record_failure(self, server_address):
-        self.history[server_address].append(InterceptionResult.failure)
+        self.history[server_address].append(InterceptionResult.FAILURE)
 
     def record_skipped(self, server_address):
-        self.history[server_address].append(InterceptionResult.skipped)
+        self.history[server_address].append(InterceptionResult.SKIPPED)
 
 
-class ConservativeStrategy(_TlsStrategy):
+class ConservativeStrategy(TlsStrategy):
     """
     Conservative Interception Strategy - only intercept if there haven't been any failed attempts
     in the history.
     """
-
-    def should_intercept(self, server_address):
-        if InterceptionResult.failure in self.history[server_address]:
-            return False
-        return True
+    def should_intercept(self, server_address: connection.Address) -> bool:
+        return InterceptionResult.FAILURE not in self.history[server_address]
 
 
-class ProbabilisticStrategy(_TlsStrategy):
+class ProbabilisticStrategy(TlsStrategy):
     """
     Fixed probability that we intercept a given connection.
     """
-
-    def __init__(self, p):
+    def __init__(self, p: float):
         self.p = p
         super().__init__()
 
-    def should_intercept(self, server_address):
+    def should_intercept(self, server_address: connection.Address) -> bool:
         return random.uniform(0, 1) < self.p
 
 
-class TlsFeedback(TlsLayer):
-    """
-    Monkey-patch _establish_tls_with_client to get feedback if TLS could be established
-    successfully on the client connection (which may fail due to cert pinning).
-    """
+class MaybeTls:
+    strategy: TlsStrategy
 
-    def _establish_tls_with_client(self):
-        server_address = self.server_conn.address
+    def load(self, l):
+        l.add_option(
+            "tls_strategy", int, 0,
+            "TLS passthrough strategy. If set to 0, connections will be passed through after the first unsuccessful "
+            "handshake. If set to 0 < p <= 100, connections with be passed through with probability p.",
+        )
 
-        try:
-            super()._establish_tls_with_client()
-        except TlsProtocolException as e:
-            tls_strategy.record_failure(server_address)
-            raise e
+    def configure(self, updated):
+        if "tls_strategy" not in updated:
+            return
+        if ctx.options.tls_strategy > 0:
+            self.strategy = ProbabilisticStrategy(ctx.options.tls_strategy / 100)
         else:
-            tls_strategy.record_success(server_address)
+            self.strategy = ConservativeStrategy()
 
+    def tls_clienthello(self, data: tls.ClientHelloData):
+        server_address = data.context.server.peername
+        if not self.strategy.should_intercept(server_address):
+            ctx.log(f"TLS passthrough: {human.format_address(server_address)}.")
+            data.ignore_connection = True
+            self.strategy.record_skipped(server_address)
 
-# inline script hooks below.
-
-tls_strategy = None
-
-
-def load(l):
-    l.add_option(
-        "tlsstrat", int, 0, "TLS passthrough strategy (0-100)",
-    )
-
-
-def configure(updated):
-    global tls_strategy
-    if ctx.options.tlsstrat > 0:
-        tls_strategy = ProbabilisticStrategy(float(ctx.options.tlsstrat) / 100.0)
-    else:
-        tls_strategy = ConservativeStrategy()
-
-
-def next_layer(next_layer):
-    """
-    This hook does the actual magic - if the next layer is planned to be a TLS layer,
-    we check if we want to enter pass-through mode instead.
-    """
-    if isinstance(next_layer, TlsLayer) and next_layer._client_tls:
-        server_address = next_layer.server_conn.address
-
-        if tls_strategy.should_intercept(server_address):
-            # We try to intercept.
-            # Monkey-Patch the layer to get feedback from the TLSLayer if interception worked.
-            next_layer.__class__ = TlsFeedback
+    def tls_handshake(self, data: tls.TlsHookData):
+        if isinstance(data.conn, connection.Server):
+            return
+        server_address = data.context.server.peername
+        if data.conn.error is None:
+            ctx.log(f"TLS handshake successful: {human.format_address(server_address)}")
+            self.strategy.record_success(server_address)
         else:
-            # We don't intercept - reply with a pass-through layer and add a "skipped" entry.
-            mitmproxy.ctx.log("TLS passthrough for %s" % repr(next_layer.server_conn.address), "info")
-            next_layer_replacement = RawTCPLayer(next_layer.ctx, ignore=True)
-            next_layer.reply.send(next_layer_replacement)
-            tls_strategy.record_skipped(server_address)
+            ctx.log(f"TLS handshake failed: {human.format_address(server_address)}")
+            self.strategy.record_failure(server_address)
+
+
+addons = [MaybeTls()]